requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==5.0.4
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-filter==24.2
django-rest-framework==0.1.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
environs==11.0.0
fire==0.6.0
flake8==7.0.0
fonttools==4.51.0
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.6
lxml==5.2.1
marshmallow==3.21.1
mccabe==0.7.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.3.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
six==1.16.0
sqlparse==0.4.4
termcolor==2.4.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
